+++ Package.swift
@@ -3,12 +3,13 @@ import PackageDescription
 
 let package = Package(
     name: "TapTagKit",
+    defaultLocalization: "en",
     platforms: [.iOS(.v15)],
     products: [
         .library(name: "TapTagKit", targets: ["TapTagKit"]),
     ],
     targets: [
-        .target(name: "TapTagKit"),
+        .target(name: "TapTagKit", resources: [.process("Resources")]),
         .testTarget(name: "TapTagKitTests", dependencies: ["TapTagKit"]),
     ]
 )